+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Fix usage of `:in` when defined with a `:default` option that is not a hash, to apply the default when param is not given.
+* Fix usage of `:in` incorrectly calling scopes when receiving a blank param value without `allow_blank` set.
 * Require `active_support` and `action_controller` explicitly to prevent possible `uninitialized constant` errors.
 * Add support for Ruby 2.7. (no changes required)
 * Remove support for Rails `4.1` and `4.2` and Ruby `2.1`, `2.2`, and `2.3`.