+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Fix usage of `:in` when defined with a `:default` option that is not a hash, to apply the default when param is not given.
+* Fix usage of `:in` incorrectly calling scopes when receiving a blank param value without `allow_blank` set.
 * Deprecate passing a String to `if` and `unless` options, in order to stop using `eval` in code.
 * Require `active_support` and `action_controller` explicitly to prevent possible `uninitialized constant` errors.
 * Add support for Ruby 2.7 and 3.0, drop support for Ruby < 2.5.